dayum, missed one

+++ crates/re_arrow_store/src/store_write.rs
@@ -255,7 +255,7 @@ impl ComponentTable {
         let size = bucket.total_size_bytes();
         let size_overflow = bucket.total_size_bytes() >= config.component_bucket_size_bytes;
 
-        let len = bucket.total_rows() as u64;
+        let len = bucket.total_rows();
         let len_overflow = len >= config.component_bucket_nb_rows;
 
         if size_overflow || len_overflow {
